+++ src/server/world/palette.h
@@ -25,35 +25,35 @@
 
 class Palette {
  public:
-  Palette(int, std::set<unsigned short>);
+  Palette(int, std::set<uint16_t>);
   virtual ~Palette();
 
  public:
-  virtual std::vector<unsigned long long> apply(std::vector<unsigned short>) = 0;
+  virtual std::vector<uint64_t> apply(std::vector<uint16_t>) = 0;
   virtual void write(ByteBuffer&) = 0;
 
  public:
   int bitsPerEntry;
-  std::set<unsigned short> palette;
+  std::set<uint16_t> palette;
 };
 
 class SingleValuedPalette : public Palette {
  public:
-  explicit SingleValuedPalette(std::set<unsigned short>);
+  explicit SingleValuedPalette(std::set<uint16_t>);
   ~SingleValuedPalette() override;
 
  public:
-  std::vector<unsigned long long> apply(std::vector<unsigned short>) override;
+  std::vector<uint64_t> apply(std::vector<uint16_t>) override;
   void write(ByteBuffer&) override;
 };
 
 class IndirectPalette : public Palette {
  public:
-  IndirectPalette(int, std::set<unsigned short>);
+  IndirectPalette(int, std::set<uint16_t>);
   ~IndirectPalette() override;
 
  public:
-  std::vector<unsigned long long> apply(std::vector<unsigned short>) override;
+  std::vector<uint64_t> apply(std::vector<uint16_t>) override;
   void write(ByteBuffer&) override;
 };
 
@@ -63,6 +63,6 @@ class DirectPalette : public Palette {
   ~DirectPalette() override;
 
  public:
-  std::vector<unsigned long long> apply(std::vector<unsigned short>) override;
+  std::vector<uint64_t> apply(std::vector<uint16_t>) override;
   void write(ByteBuffer&) override;
 };
